Kinda organize navigation related controls
