+++ src/Wpf.Ui/Controls/Navigation/NavigationBackButton.cs
@@ -1,4 +1,10 @@
-﻿#nullable enable
+﻿// This Source Code Form is subject to the terms of the MIT License.
+// If a copy of the MIT was not distributed with this file, You can obtain one at https://opensource.org/licenses/MIT.
+// Copyright (C) Leszek Pomianowski and WPF UI Contributors.
+// All Rights Reserved.
+
+#nullable enable
+
 using System.ComponentModel;
 using System.Windows;
 using Wpf.Ui.Controls.Interfaces;
@@ -6,7 +12,7 @@ using Wpf.Ui.Controls.Interfaces;
 namespace Wpf.Ui.Controls.Navigation;
 
 /// <summary>
-/// Inherited from the <see cref="System.Windows.Controls.Button"/>, used to navigate backwards/>.
+/// Inherited from the <see cref="System.Windows.Controls.Button"/>, used to navigate backwards inside the <see cref="INavigation"/>.
 /// </summary>
 public class NavigationBackButton : System.Windows.Controls.Button
 {
@@ -17,7 +23,7 @@ public class NavigationBackButton : System.Windows.Controls.Button
         typeof(INavigation), typeof(NavigationBackButton), new PropertyMetadata(null));
 
     /// <summary>
-    /// <see cref="INavigation"/>
+    /// Parent <see cref="INavigation"/> control.
     /// </summary>
     [Bindable(true), Category("Behavior")]
     public INavigation? Navigation
@@ -28,6 +34,6 @@ public class NavigationBackButton : System.Windows.Controls.Button
 
     public NavigationBackButton()
     {
-        SetValue(CommandProperty, new Common.RelayCommand(o => Navigation?.NavigateBack(), () => Navigation is not null && Navigation.CanGoBack));
+        SetValue(CommandProperty, new Common.RelayCommand(_ => Navigation?.NavigateBack(), () => Navigation is not null && Navigation.CanGoBack));
     }
 }
