+++ src/Wpf.Ui/Controls/Navigation/NavigationHeader.bmp
