+++ src/Wpf.Ui/Controls/Navigation/NavigationSeparator.bmp
