+++ src/Wpf.Ui/Controls/Navigation/NavigationSeparator.cs
@@ -7,7 +7,7 @@ using System.ComponentModel;
 using System.Drawing;
 using Wpf.Ui.Controls.Interfaces;
 
-namespace Wpf.Ui.Controls;
+namespace Wpf.Ui.Controls.Navigation;
 
 [ToolboxItem(true)]
 [ToolboxBitmap(typeof(NavigationSeparator), "NavigationSeparator.bmp")]
