+++ src/Wpf.Ui/Styles/Controls/NavigationBackButton.xaml
@@ -2,9 +2,9 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="clr-namespace:Wpf.Ui.Controls"
-    xmlns:ui="clr-namespace:Wpf.Ui.Controls.Navigation">
+    xmlns:navigation="clr-namespace:Wpf.Ui.Controls.Navigation">
 
-    <Style TargetType="{x:Type ui:NavigationBackButton}">
+    <Style TargetType="{x:Type navigation:NavigationBackButton}">
         <Setter Property="Foreground">
             <Setter.Value>
                 <SolidColorBrush Color="{DynamicResource TextFillColorSecondary}" />
@@ -24,7 +24,7 @@
         <Setter Property="OverridesDefaultStyle" Value="True" />
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type ui:NavigationBackButton}">
+                <ControlTemplate TargetType="{x:Type navigation:NavigationBackButton}">
                     <Border
                         x:Name="MainBorder"
                         Margin="0"
@@ -64,7 +64,7 @@
                                         Storyboard.TargetName="IconScaleTransform"
                                         Storyboard.TargetProperty="(ScaleTransform.ScaleX)"
                                         From="1.0"
-                                        To="0.6"
+                                        To="0.9"
                                         Duration="0:0:0.08" />
                                 </Storyboard>
                             </BeginStoryboard>
@@ -75,7 +75,7 @@
                                     <DoubleAnimation
                                         Storyboard.TargetName="IconScaleTransform"
                                         Storyboard.TargetProperty="(ScaleTransform.ScaleX)"
-                                        From="0.6"
+                                        From="0.9"
                                         To="1.0"
                                         Duration="0:0:0.08" />
                                 </Storyboard>
