+++ src/Wpf.Ui/Styles/Controls/NavigationCompact.xaml
@@ -8,7 +8,8 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="clr-namespace:Wpf.Ui.Controls">
+    xmlns:controls="clr-namespace:Wpf.Ui.Controls"
+    xmlns:navigation="clr-namespace:Wpf.Ui.Controls.Navigation">
 
     <Style x:Key="HamburgerButton" TargetType="{x:Type controls:Button}">
         <Setter Property="Foreground">
@@ -96,7 +97,7 @@
         </Setter>
     </Style>
 
-    <Style x:Key="NavigationCompactHeader" TargetType="{x:Type controls:NavigationHeader}">
+    <Style x:Key="NavigationCompactHeader" TargetType="{x:Type navigation:NavigationHeader}">
         <Setter Property="Foreground">
             <Setter.Value>
                 <SolidColorBrush Color="{DynamicResource TextFillColorTertiary}" />
@@ -118,7 +119,7 @@
         <Setter Property="OverridesDefaultStyle" Value="True" />
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type controls:NavigationHeader}">
+                <ControlTemplate TargetType="{x:Type navigation:NavigationHeader}">
                     <Grid
                         Margin="{TemplateBinding Margin}"
                         HorizontalAlignment="{TemplateBinding HorizontalAlignment}"
@@ -400,7 +401,7 @@
         </Setter>
         <Style.Resources>
             <Style BasedOn="{StaticResource NavigationCompactItem}" TargetType="{x:Type controls:NavigationItem}" />
-            <Style BasedOn="{StaticResource NavigationCompactHeader}" TargetType="{x:Type controls:NavigationHeader}" />
+            <Style BasedOn="{StaticResource NavigationCompactHeader}" TargetType="{x:Type navigation:NavigationHeader}" />
         </Style.Resources>
     </Style>
