+++ src/Wpf.Ui/Styles/Controls/NavigationHeader.xaml
@@ -8,9 +8,10 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="clr-namespace:Wpf.Ui.Controls">
+    xmlns:controls="clr-namespace:Wpf.Ui.Controls"
+    xmlns:navigation="clr-namespace:Wpf.Ui.Controls.Navigation">
 
-    <Style TargetType="{x:Type controls:NavigationHeader}">
+    <Style TargetType="{x:Type navigation:NavigationHeader}">
         <Setter Property="Foreground">
             <Setter.Value>
                 <SolidColorBrush Color="{DynamicResource TextFillColorSecondary}" />
@@ -31,7 +32,7 @@
         <Setter Property="OverridesDefaultStyle" Value="True" />
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type controls:NavigationHeader}">
+                <ControlTemplate TargetType="{x:Type navigation:NavigationHeader}">
                     <Grid
                         Margin="{TemplateBinding Margin}"
                         HorizontalAlignment="{TemplateBinding HorizontalAlignment}"
