+++ src/Wpf.Ui/Styles/Controls/NavigationSeparator.xaml
@@ -8,9 +8,10 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="clr-namespace:Wpf.Ui.Controls">
+    xmlns:controls="clr-namespace:Wpf.Ui.Controls"
+    xmlns:navigation="clr-namespace:Wpf.Ui.Controls.Navigation">
 
-    <Style TargetType="{x:Type controls:NavigationSeparator}">
+    <Style TargetType="{x:Type navigation:NavigationSeparator}">
         <Setter Property="BorderBrush" Value="{DynamicResource ControlElevationBorderBrush}" />
         <Setter Property="Background" Value="Transparent" />
         <Setter Property="Margin" Value="0,4" />
@@ -19,7 +20,7 @@
         <Setter Property="OverridesDefaultStyle" Value="True" />
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type controls:NavigationSeparator}">
+                <ControlTemplate TargetType="{x:Type navigation:NavigationSeparator}">
                     <Border
                         Width="{TemplateBinding Width}"
                         Margin="{TemplateBinding Margin}"
