+++ src/Wpf.Ui/Styles/Controls/NavigationStore.xaml
@@ -9,9 +9,10 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:common="clr-namespace:Wpf.Ui.Common"
-    xmlns:controls="clr-namespace:Wpf.Ui.Controls">
+    xmlns:controls="clr-namespace:Wpf.Ui.Controls"
+    xmlns:navigation="clr-namespace:Wpf.Ui.Controls.Navigation">
 
-    <Style x:Key="NavigationStoreHeader" TargetType="{x:Type controls:NavigationHeader}">
+    <Style x:Key="NavigationStoreHeader" TargetType="{x:Type navigation:NavigationHeader}">
         <Setter Property="Foreground">
             <Setter.Value>
                 <SolidColorBrush Color="{DynamicResource TextFillColorTertiary}" />
@@ -33,7 +34,7 @@
         <Setter Property="OverridesDefaultStyle" Value="True" />
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type controls:NavigationHeader}">
+                <ControlTemplate TargetType="{x:Type navigation:NavigationHeader}">
                     <Grid
                         Margin="{TemplateBinding Margin}"
                         HorizontalAlignment="{TemplateBinding HorizontalAlignment}"
@@ -346,7 +347,7 @@
         </Setter>
         <Style.Resources>
             <Style BasedOn="{StaticResource NavigationStoreItem}" TargetType="{x:Type controls:NavigationItem}" />
-            <Style BasedOn="{StaticResource NavigationStoreHeader}" TargetType="{x:Type controls:NavigationHeader}" />
+            <Style BasedOn="{StaticResource NavigationStoreHeader}" TargetType="{x:Type navigation:NavigationHeader}" />
         </Style.Resources>
     </Style>
