+++ src/Wpf.Ui/Wpf.Ui.csproj
@@ -121,9 +121,9 @@
     <EmbeddedResource Include="Controls\Dialog.bmp" />
     <EmbeddedResource Include="Controls\FontIcon.bmp" />
     <EmbeddedResource Include="Controls\NavigationCompact.bmp" />
-    <EmbeddedResource Include="Controls\NavigationSeparator.bmp" />
+    <EmbeddedResource Include="Controls\Navigation\NavigationSeparator.bmp" />
     <EmbeddedResource Include="Controls\NavigationItem.bmp" />
-    <EmbeddedResource Include="Controls\NavigationHeader.bmp" />
+    <EmbeddedResource Include="Controls\Navigation\NavigationHeader.bmp" />
     <EmbeddedResource Include="Controls\NavigationFluent.bmp" />
     <EmbeddedResource Include="Controls\NavigationStore.bmp" />
     <EmbeddedResource Include="Controls\NotifyIcon.bmp" />
